[refact] Adjust App code for 'Social Login' environment * Main   * 'restorePreviousGoogleSignIn' 함수잠금   * 구글로그인 시 유저정보 확인을 위해, 이전로그인 복구기능 잠금처리   * 'Archive/Authenticator.swift' 파일로 이동 * Util   * 파일 관리 및 코드 가독성을 위한 'Utilities' 클래스 조정   * 클래스내 'topViewController' 새로 추가한 'GoogleLoginHelper' 클래스로 이동   * 기존에 발생중인 'keyWindow was deprecated in iOS 13.0' 메세지 제거를 위해 'topViewController' 업데이트     * 'UIWindowScene'을 사용하는 방식으로 변경   * 영향을 받는 'Authentication' 코드수정
